Removed async loading of CSS in favor of blocking
BAB used to load CSS in a non-blocking method. This turned out to cause issues when the JS loaded before the CSS had fully defined the size of the banner. In some instances, this caused positioning problems of elements. BAB now loads the style.css in a "blocking" fashion instead. Having your style defined is probably a requirement since this is an animated ad.